renamed pipe to pipeline, adding pipeline
