+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.0.0] - UNRELEASED
+### Renamed
+ - pipe to pipeline
+
+### Added
+ - pipe
+
 ## [6.2.3] - 2019-1-23
 ### Added
  -  flat, asyncFlat, cursor, asyncCursor, merge, asyncMerge, pipe, partition, async-partition
