+++ src/pipeline.d.ts
@@ -1,0 +1,2 @@
+import { pipeline } from './index'
+export default pipeline
