+++ src/pipeline.mjs
@@ -1,0 +1,3 @@
+export default function pipeline (value, ...fns) {
+  return fns.reduce((value, fn) => fn(value), value)
+}
